fix: <double> para <int>

+++ mergeSort/mergeSort.h
@@ -2,5 +2,5 @@
 #include <vector>
 using namespace std;
 
-void mergeSort(vector <double> &vetor, int inicio, int fim, int meio, int &contador);
-void merge(vector <double> &vetor,int &contador);
+void mergeSort(vector <int> &vetor, int inicio, int fim, int meio, int &contador);
+void merge(vector <int> &vetor,int &contador);
